+++ lib/host_kit/local.ex
@@ -175,6 +175,10 @@ defmodule HostKit.Local do
     Helpers.read_env_file(desired, &read(&1, context))
   end
 
+  def read(%HostKit.Resources.ConfigFile{} = desired, context) do
+    Helpers.read_config_file(desired, &read(&1, context))
+  end
+
   def read(%HostKit.Resources.Template{} = desired, context) do
     Helpers.read_template(desired, &read(&1, context))
   end